adds new popup structure
