+++ src/presenter/popup-presenter.js
@@ -1,5 +1,11 @@
 import { isEscapeEvent } from '../util/util.js';
 import { render, remove, replace } from '../framework/render.js';
+import CommentsContainerView from '../view/comments-container-view.js';
+import CommentsSectionView from '../view/comments-section-view.js';
+import CommentsTitleView from '../view/comments-title-view.js';
+import CommentsPresenter from './comments-presenter.js';
+import CommentsNewPresenter from './comment-new-presenter.js';
+import FilmDetailPresenter from './film-details-presenter.js';
 import PopupView from '../view/popup-view.js';
 
 const footerElement = document.querySelector('.footer');
@@ -12,6 +18,10 @@ export default class PopupPresenter {
   #popupComponent = null;
   #container = footerElement;
 
+  #commentsPresenter;
+  #commentsNewPresenter;
+  #filmDetailPresenter;
+
   constructor(changeData) {
     this.#changeData = changeData;
   }
@@ -22,25 +32,23 @@ export default class PopupPresenter {
 
     const prevPopupComponent = this.#popupComponent;
 
-    this.#popupComponent = new PopupView(this.#film, this.#getFilmComments());
+    this.#popupComponent = new PopupView();
 
     if (prevPopupComponent === null) {
       this.#renderPopup();
+      this.#renderFilmDetails();
+      this.#renderComments();
       this.#setupFilmUserDetailHandlers();
-
-      document.body.classList.add('hide-overflow');
-      document.addEventListener('keydown', this.#onEscKeyDown);
-      this.#popupComponent.setClickHandler(this.#onFilmDetailsCloseBtnClick);
+      this.#setupCloseHandlers();
       return;
     }
 
     if (bodyElement.contains(prevPopupComponent.element)) {
+      this.#renderFilmDetails();
+      this.#renderComments();
       replace(this.#popupComponent, prevPopupComponent);
       this.#setupFilmUserDetailHandlers();
-
-      document.body.classList.add('hide-overflow');
-      document.addEventListener('keydown', this.#onEscKeyDown);
-      this.#popupComponent.setClickHandler(this.#onFilmDetailsCloseBtnClick);
+      this.#setupCloseHandlers();
     }
 
     remove(prevPopupComponent);
@@ -53,17 +61,44 @@ export default class PopupPresenter {
 
   #getFilmComments = () => this.#film.comments.map((item) => this.#commentsModel.getCommentById(item));
 
+  #renderFilmDetails = () => {
+    this.#filmDetailPresenter = new FilmDetailPresenter(this.#film, this.#popupComponent);
+    this.#filmDetailPresenter.init();
+  };
+
+  #renderComments = () => {
+    this.commentsContainerComponent = new CommentsContainerView();
+    render(this.commentsContainerComponent, this.#popupComponent.element);
+    this.commentsSectionComponent = new CommentsSectionView();
+    render(this.commentsSectionComponent, this.commentsContainerComponent.element);
+    render(new CommentsTitleView(this.#commentsModel.comments), this.commentsSectionComponent.element);
+
+    this.#commentsPresenter = new CommentsPresenter(this.#getFilmComments(), this.commentsSectionComponent);
+    this.#commentsPresenter.init();
+
+    this.#commentsNewPresenter = new CommentsNewPresenter(this.commentsSectionComponent);
+    this.#commentsNewPresenter.init();
+  };
+
   closePopup = () => {
     remove(this.#popupComponent);
     this.#popupComponent = null;
     document.body.classList.remove('hide-overflow');
     document.removeEventListener('keydown', this.#onEscKeyDown);
+    //clear input and object
+  };
+
+  #setupCloseHandlers = () => {
+    document.body.classList.add('hide-overflow');
+    document.addEventListener('keydown', this.#onEscKeyDown);
+    this.#filmDetailPresenter.getFilmDetailComponent().setClickHandler(this.#onFilmDetailsCloseBtnClick);
   };
 
   #setupFilmUserDetailHandlers = () => {
-    this.#popupComponent.setWatchlistClickHandler(this.#handleWatchlistClick);
-    this.#popupComponent.setAlreadyWatchedClickHandler(this.#handleAlreadyWatchedClick);
-    this.#popupComponent.setFavoriteClickHandler(this.#handleFavoriteClick);
+    const filmDetailComponent = this.#filmDetailPresenter.getFilmDetailComponent();
+    filmDetailComponent.setWatchlistClickHandler(this.#handleWatchlistClick);
+    filmDetailComponent.setAlreadyWatchedClickHandler(this.#handleAlreadyWatchedClick);
+    filmDetailComponent.setFavoriteClickHandler(this.#handleFavoriteClick);
   };
 
   #handleWatchlistClick = () => {
