+++ src/view/film-details-view.js
@@ -1,0 +1,165 @@
+import AbstractView from '../framework/view/abstract-view.js';
+import { transformIntToHour, humanizeDateRelease } from '../util/util.js';
+
+const createGenresTemplate = (genres) => genres.map((item) => `<span class="film-details__genre">${item}</span>`);
+
+const createFilmDetailsTemplate = (film) => {
+  const {
+    filmInfo: {
+      title,
+      alternativeTitle,
+      totalRating,
+      poster,
+      ageRating,
+      director,
+      writers,
+      actors,
+      release: {
+        date,
+        releaseCountry
+      },
+      runtime,
+      genre,
+      description,
+    },
+    userDetails: {
+      watchList,
+      alreadyWatched,
+      favorite,
+    },
+  } = film;
+
+  const watchListActive = watchList
+    ? 'film-details__control-button--active'
+    : '';
+  const alreadyWatchedActive = alreadyWatched
+    ? 'film-details__control-button--active'
+    : '';
+  const favoriteActive = favorite
+    ? 'film-details__control-button--active'
+    : '';
+  const genresTemplate = createGenresTemplate(genre);
+
+  return `<div class="film-details__top-container">
+    <div class="film-details__close">
+      <button class="film-details__close-btn" type="button">close</button>
+    </div>
+    <div class="film-details__info-wrap">
+      <div class="film-details__poster">
+        <img class="film-details__poster-img" src="${poster}" alt="">
+
+        <p class="film-details__age">${ageRating}+</p>
+      </div>
+
+      <div class="film-details__info">
+        <div class="film-details__info-head">
+          <div class="film-details__title-wrap">
+            <h3 class="film-details__title">${title}</h3>
+            <p class="film-details__title-original">Original: ${alternativeTitle}</p>
+          </div>
+
+          <div class="film-details__rating">
+            <p class="film-details__total-rating">${totalRating}</p>
+          </div>
+        </div>
+
+        <table class="film-details__table">
+          <tbody><tr class="film-details__row">
+            <td class="film-details__term">Director</td>
+            <td class="film-details__cell">${director}</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Writers</td>
+            <td class="film-details__cell">${writers.join(', ').trim()}</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Actors</td>
+            <td class="film-details__cell">${actors.join(', ').trim()}</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Release Date</td>
+            <td class="film-details__cell">${humanizeDateRelease(date)}</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Runtime</td>
+            <td class="film-details__cell">${transformIntToHour(runtime)}</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Country</td>
+            <td class="film-details__cell">${releaseCountry}</td>
+          </tr>
+          <tr class="film-details__row">
+            <td class="film-details__term">Genres</td>
+            <td class="film-details__cell">
+            ${genresTemplate}
+            </td>
+          </tr>
+        </tbody></table>
+
+        <p class="film-details__film-description">
+        ${description}
+        </p>
+      </div>
+    </div>
+
+    <section class="film-details__controls">
+      <button type="button" class="film-details__control-button film-details__control-button--watchlist ${watchListActive}" id="watchlist" name="watchlist">Add to watchlist</button>
+      <button type="button" class="film-details__control-button  film-details__control-button--watched ${alreadyWatchedActive}" id="watched" name="watched">Already watched</button>
+      <button type="button" class="film-details__control-button film-details__control-button--favorite ${favoriteActive}" id="favorite" name="favorite">Add to favorites</button>
+    </section>
+  </div>`;
+};
+
+export default class FilmDetailsView extends AbstractView {
+  #film;
+
+  constructor(film) {
+    super();
+    this.#film = film;
+  }
+
+  get template() {
+    return createFilmDetailsTemplate(this.#film);
+  }
+
+  setClickHandler = (callback) => {
+    this._callback.click = callback;
+    this.element.querySelector('.film-details__close-btn').addEventListener('click', this.#clickHandler);
+  };
+
+  setWatchlistClickHandler = (callback) => {
+    this._callback.watchlistClick = callback;
+    this.element.querySelector('.film-details__control-button--watchlist').addEventListener('click', this.#watchlistClickHandler);
+  };
+
+  setAlreadyWatchedClickHandler = (callback) => {
+    this._callback.alreadyWatchedClick = callback;
+    this.element.querySelector('.film-details__control-button--watched').addEventListener('click', this.#alreadyWatchedClickHandler);
+  };
+
+  setFavoriteClickHandler = (callback) => {
+    this._callback.favoriteClick = callback;
+    this.element.querySelector('.film-details__control-button--favorite').addEventListener('click', this.#favoriteClickHandler);
+  };
+
+  #clickHandler = (evt) => {
+    evt.preventDefault();
+    this._callback.click();
+  };
+
+  #watchlistClickHandler = (evt) => {
+    evt.preventDefault();
+    this._callback.watchlistClick();
+  };
+
+  #alreadyWatchedClickHandler = (evt) => {
+    evt.preventDefault();
+    this._callback.alreadyWatchedClick();
+  };
+
+  #favoriteClickHandler = (evt) => {
+    evt.preventDefault();
+    this._callback.favoriteClick();
+  };
+
+}
